# src/openapi.yaml
openapi: 3.1.0
info:
  title: MCP REST Bridge
  version: 1.0.0
  description: >
    A self-hosted MCP gateway that bridges MCP clients to backend REST APIs and
    other MCP servers, bidirectionally. REST-capable services register
    themselves (via OpenAPI/Swagger auto-discovery, a cURL/Postman import, or a
    manual tool list) and the bridge exposes their tools to any MCP client
    (e.g. Claude Desktop, Cursor, or a custom agent) via the Streamable HTTP
    MCP transport; existing MCP servers can likewise be registered as
    upstreams and re-exposed. Every call is proxied through a uniform guard
    pipeline (SSRF check, guardrails, per-tool policy, circuit breaker,
    response sanitizing, audit) documented in this admin API — this
    specification covers that management surface, not the dynamic per-client
    tool endpoints it provisions.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /admin-api/alerts:
    get:
      operationId: listAlertRules
      summary: List webhook alert rules
      tags: [Alerts]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      responses:
        "200":
          description: All alert rules.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      properties: { items: { type: array, items: { $ref: "#/components/schemas/AlertRule" } } },
                    },
                },
            }
    post:
      operationId: createAlertRule
      summary: Create a new alert rule
      description: Requires the admin role for session callers.
      tags: [Alerts]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, eventType, webhookUrl]
              properties:
                name: { type: string, maxLength: 128 }
                eventType:
                  {
                    type: string,
                    enum: [circuit_breaker_open, client_unreachable, error_rate, usage_spike, schema_drift],
                  }
                webhookUrl: { type: string, description: Absolute http(s) URL. }
                threshold:
                  { type: [number, "null"], description: "Meaning depends on eventType, e.g. an error-rate fraction." }
                minCalls:
                  {
                    type: [number, "null"],
                    description: Minimum sample size before an error_rate/usage_spike rule can fire.,
                  }
      responses:
        "201":
          description: Created.
          content: { application/json: { schema: { $ref: "#/components/schemas/AlertRule" } } }
        "400":
          description: Invalid name/eventType/webhookUrl/threshold/minCalls.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin-api/alerts/{id}:
    patch:
      operationId: updateAlertRule
      summary: Update an alert rule
      description: Requires the admin role for session callers.
      tags: [Alerts]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                enabled: { type: boolean }
                webhookUrl: { type: string }
                threshold: { type: [number, "null"] }
                minCalls: { type: [number, "null"] }
      responses:
        "200":
          description: Updated.
          content: { application/json: { schema: { $ref: "#/components/schemas/AlertRule" } } }
        "400":
          description: Invalid field value.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: No such alert rule.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    delete:
      operationId: deleteAlertRule
      summary: Delete an alert rule
      description: Requires the admin role for session callers.
      tags: [Alerts]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted.
          content:
            {
              application/json:
                {
                  schema:
                    { type: object, properties: { status: { type: string, example: deleted }, id: { type: integer } } },
                },
            }
        "404":
          description: No such alert rule.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
...
components:
...
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      description: >
        API key for administrative endpoints (registration, introspection).
        Set via ADMIN_API_KEYS env var. Disabled when AUTH_DISABLED=true.
    McpAuth:
      type: http
      scheme: bearer
      description: >
        API key for the backend-tool MCP transport endpoints
        (/mcp/{clientName}, /mcp-custom/{bundleName}). Set via MCP_API_KEYS
        env var, or a managed key (see MCP Keys). Disabled when
        AUTH_DISABLED=true or when no keys are configured at all.
    RootMcpAuth:
      type: http
      scheme: bearer
      description: >
        Credential for the /mcp system control-plane. Either the env admin
        Bearer key (ADMIN_API_KEYS) or a managed MCP key with `adminRole`
        set (see MCP Keys). Unlike McpAuth, there is no "no auth material
        configured => allow all" fallback — a caller must resolve to a real
        system role or the request is rejected outright, even on a fresh
        install with zero keys minted. Still bypassed entirely when
        AUTH_DISABLED=true (the same global dev/test escape hatch every
        other auth path in this API respects).
    SessionAuth:
      type: apiKey
      in: cookie
      name: __Host-mcp_admin_session
      description: >
        Admin-UI session cookie, issued by POST /admin-api/auth/login.
        Accepted anywhere AdminAuth is accepted. On mutating requests
        (anything but GET/HEAD/OPTIONS) an `X-CSRF-Token` header matching the
        `__Host-mcp_admin_csrf` cookie value is also required — OpenAPI 3.1
        has no first-class way to express that pairing, so it's documented
        here in prose, mirroring how RegistrationPayload's tools/openapi_url
        mutual-exclusivity is documented rather than schema-encoded.
        The `__Host-` prefix (and therefore this exact name) only applies
        when SESSION_COOKIE_SECURE=true (the default) — it mandates the
...
  schemas:
...
    AlertRule:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        eventType:
          { type: string, enum: [circuit_breaker_open, client_unreachable, error_rate, usage_spike, schema_drift] }
        enabled: { type: boolean }
        webhookUrl: { type: string }
        threshold: { type: [number, "null"] }
        minCalls: { type: [number, "null"] }
        lastFiredAt: { type: [integer, "null"], description: Unix ms epoch }
        createdAt: { type: integer, description: Unix ms epoch }
        updatedAt: { type: integer, description: Unix ms epoch }
        createdBy: { type: [string, "null"] }

    # -------------------------------------------------------------------------
    # Schedules
    # -------------------------------------------------------------------------
...
    ApiError:
      type: object
      description: >
        Standard error envelope returned by the bridge's REST endpoints (not
        MCP transport endpoints, which use JSON-RPC 2.0 error objects instead).
      required:
        - error
      properties:
        error:
          type: object
          description: Container for the structured error detail.
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: >
                Machine-readable error code in SCREAMING_SNAKE_CASE.
                Known values: VALIDATION_ERROR, CLIENT_NOT_FOUND, DISCOVERY_ERROR,
                UNAUTHORIZED, FORBIDDEN, RATE_LIMITED.
              example: VALIDATION_ERROR
            message:
              type: string
              description: >
                Human-readable description of what went wrong, suitable for
                displaying in a developer console or log.
              example: "Missing required fields: name, health_url"
            retry_after:
              type: integer
              description: >
                Seconds the caller should wait before retrying. Present only on
                a 429 RATE_LIMITED response (also echoed in the Retry-After
                header); omitted on every other error.
              example: 45
            request_id:
              type: [string, "null"]
              description: >
                Correlation id for this request (from res.locals.requestId),
                echoed back so a caller can quote it when reporting an issue.
                Present on almost every error response in the admin API;
                absent (or omitted) on the handful of handlers that don't
                thread it through (notably GET /clients and GET
                /clients/{name}/tools's error bodies, and /metrics's disabled
                response).
              example: "3f9c9a2e-9c34-4b3a-9a0e-4f7a6f6c9a1b"
      example:
        error:
          code: VALIDATION_ERROR
          message: "Missing required fields: name, health_url"
          request_id: "3f9c9a2e-9c34-4b3a-9a0e-4f7a6f6c9a1b"

    # -------------------------------------------------------------------------
    # JsonRpcMessage  (internal — MCP transport)
    # -------------------------------------------------------------------------
...